move db.server.ts from utils to config
